using shared data
Shared data can be used across the app and it allows to save the data
when we want to terminate the program instead of saving it every time
the word is learned.

An elegant solution is using Arc and Mutex due to the second thread
which is provided by ctrl+c dependency.